feat: change onChange callback to use event instead of string
BREAKING CHANGE: This updates the onChange callback parameter from a string to an event. This change
is necessary for integrating the NumericFormat component, as it expects an event object. While this
does introduce a breaking change, it offers greater flexibility and aligns more closely with the
common patterns for handling events in React.