feat: allow customization of replaced decimal when typing
When typing a comma as decimal separator, it's immediatly replaced with a dot. The NumericField
component API has been updated according to the Maskito library API. This commit also fixes jumping
cursor to the left of the default zero value when the user removes the content of the field. This
issue was caused by the `react-number-format` library. Since we store numbers in the Slice, we need
to prevent to lift up intermediate value.